Make hamming_weight(bytea) returns bigint.
This is required to work with large bytea values (up to 1GB).

+++ hamming_weight/hamming_weight.sql
@@ -2,7 +2,7 @@
   Compute the Hamming weight (count of bits set to '1')
   for bytea, int4, int8
 */
-CREATE FUNCTION hamming_weight(bytea) RETURNS int
+CREATE FUNCTION hamming_weight(bytea) RETURNS bigint
  AS '$libdir/hamming_weight.so', 'hamming_weight_bytea'
 LANGUAGE C immutable strict;
